Created and integrated position data package
