+++ backend/session/data/participant_data.py
@@ -1,7 +1,7 @@
 from dataclasses import dataclass, field
 
 from session.data.base_data import BaseData
-from session.data.position_data import PositionData
+from session.data.position.position_data import PositionData
 from session.data.size.size_data import SizeData
